requirements.txt
black
dash
dash-bootstrap-components
dash_daq
flake8
python-dotenv
kafka-python
pandas
psycopg2-binary
sqlalchemy
requests
